Mid-Work commit. placed all the private functions of the mainViewModel into a new class . Will make them work together and then through the backgrround worker
