--- WPFHook/ViewModels/EventLoop.cs
@@ -1,56 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Runtime.InteropServices;
-using System.Text;
-
-namespace WPFHook.ViewModels
-{
-    public static class EventLoop
-    {
-        /*public BackgroundWorker backgroundWorker;
-        public EventLoop(BackgroundWorker backgroundWorker)
-        {
-            this.backgroundWorker = backgroundWorker;
-        }*/
-        public static void Run()
-        {
-            MSG msg;
-
-            while (true)
-            {
-                if (PeekMessage(out msg, IntPtr.Zero, 0, 0, PM_REMOVE))
-                {
-                    if (msg.Message == WM_QUIT)
-                        break;
-
-                    TranslateMessage(ref msg);
-                    DispatchMessage(ref msg);
-                }
-            }
-        }
-
-        [StructLayout(LayoutKind.Sequential)]
-        private struct MSG
-        {
-            public IntPtr Hwnd;
-            public uint Message;
-            public IntPtr WParam;
-            public IntPtr LParam;
-            public uint Time;
-            public System.Drawing.Point Point;
-        }
-
-        const uint PM_NOREMOVE = 0;
-        const uint PM_REMOVE = 1;
-
-        const uint WM_QUIT = 0x0012;
-
-        [DllImport("user32.dll")]
-        private static extern bool PeekMessage(out MSG lpMsg, IntPtr hwnd, uint wMsgFilterMin, uint wMsgFilterMax, uint wRemoveMsg);
-        [DllImport("user32.dll")]
-        private static extern bool TranslateMessage(ref MSG lpMsg);
-        [DllImport("user32.dll")]
-        private static extern IntPtr DispatchMessage(ref MSG lpMsg);
-    }
-}
